Split up existing tests
