+++ testproject/fiber_test/tests/test_models.py
@@ -1,23 +1,16 @@
 import json
 import time
 from datetime import datetime
-from mock import patch
 
-from django.conf.urls import patterns, url
-from django.views.generic import View
-from django.core.urlresolvers import reverse
-from django.template import Template, Context
-from django.core.exceptions import ValidationError
+from mock import patch
 from django.test import TestCase
 from django.conf import settings
 from django.contrib.auth.models import User
 
-from fiber.utils.validators import FiberURLValidator
-from fiber.mixins import FiberPageMixin
-from fiber.sitemaps import FiberSitemap
 from fiber.models import Page, ContentItem, PageContentItem
 from fiber.utils import date
 from ..test_util import get_short_django_version, format_list, condense_html_whitespace
+
 
 try:
     from django.utils.timezone import make_aware, utc
@@ -329,329 +322,6 @@ class PageContentItemTest(TestCase):
         self.assertEquals(u'c a', get_content(page.id, 'side'))
 
 
-class TestTemplateTags(TestCase):
-
-    def setUp(self):
-        """
-        Generates test data
-        """
-        # generate data
-        main = Page.objects.create(title='main')
-        home = Page.objects.create(title='home', parent=main, url='/')
-        section1 = Page.objects.create(title='section1', parent=home, url='section1')
-        section2 = Page.objects.create(title='section2', parent=home, url='section2')
-        section11 = Page.objects.create(title='sub1', parent=section1, url='sub1')
-        section12 = Page.objects.create(title='sub2', parent=section1, url='sub2')
-
-    def get_non_staff_user(self):
-        user = User.objects.create_user('user1', 'u@ser.nl', password="pass")
-        user.is_staff = False
-        return user
-
-    def get_staff_user(self):
-        user = User.objects.create_user('user2', 'u2@ser.nl', password="secure")
-        user.is_staff = True
-        return user
-
-    def test_show_user_menu_all(self):
-        # render menu with all pages
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 1 999 "all" %}
-            """
-        )
-        c = Context({
-            'user': self.get_non_staff_user(),
-            'fiber_page': Page.objects.get_by_url('/'),
-        })
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                   '<li class="home first last">'
-                     '<a href="/">home</a>'
-                     '<ul>'
-                       '<li class="section1 first">'
-                         '<a href="/section1/">section1</a>'
-                         '<ul>'
-                           '<li class="sub1 first">'
-                             '<a href="/section1/sub1/">sub1</a>'
-                           '</li>'
-                           '<li class="sub2 last">'
-                             '<a href="/section1/sub2/">sub2</a>'
-                           '</li>'
-                         '</ul>'
-                       '</li>'
-                       '<li class="section2 last">'
-                         '<a href="/section2/">section2</a>'
-                       '</li>'
-                     '</ul>'
-                   '</li>'
-                 '</ul>'))
-
-    def test_show_user_menu_all_descendants(self):
-        """
-        Tests for 'all_descendants' with a minimum level
-        """
-        user = self.get_non_staff_user()
-
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 2 999 "all_descendants" %}
-            """
-        )
-
-        c = Context({
-            'user': user,
-            'fiber_page': Page.objects.get_by_url('/section1/'),
-        })
-
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                   '<li class="section1 first">'
-                     '<a href="/section1/">section1</a>'
-                     '<ul>'
-                       '<li class="sub1 first">'
-                         '<a href="/section1/sub1/">sub1</a>'
-                       '</li>'
-                       '<li class="sub2 last">'
-                         '<a href="/section1/sub2/">sub2</a>'
-                       '</li>'
-                     '</ul>'
-                   '</li>'
-                   '<li class="section2 last">'
-                     '<a href="/section2/">section2</a>'
-                   '</li>'
-                 '</ul>'))
-
-        c = Context({
-            'user': user,
-            'fiber_page': Page.objects.get_by_url('/section2/'),
-        })
-
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                   '<li class="section1 first">'
-                     '<a href="/section1/">section1</a>'
-                   '</li>'
-                   '<li class="section2 last">'
-                     '<a href="/section2/">section2</a>'
-                   '</li>'
-                 '</ul>'))
-
-    def test_show_user_menu_min_max_level(self):
-        """
-        Test for minimum and maximum level
-        """
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 2 2 %}
-            """
-        )
-        c = Context({
-            'user': self.get_non_staff_user(),
-            'fiber_page': Page.objects.get_by_url('/section1/sub1/'),
-        })
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                   '<li class="section1 first">'
-                     '<a href="/section1/">section1</a>'
-                   '</li>'
-                   '<li class="section2 last">'
-                     '<a href="/section2/">section2</a>'
-                   '</li>'
-                 '</ul>'))
-
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 3 3 %}
-            """
-        )
-
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                   '<li class="sub1 first">'
-                     '<a href="/section1/sub1/">sub1</a>'
-                   '</li>'
-                   '<li class="sub2 last">'
-                     '<a href="/section1/sub2/">sub2</a>'
-                   '</li>'
-                 '</ul>'))
-
-    def test_show_user_menu_different_root(self):
-        """
-        Test that show_menu only shows top level if current
-        page is in different root.
-        """
-        other_root = Page.objects.create(title='other')
-
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 1 999 %}
-            """
-        )
-        c = Context({
-            'user': self.get_non_staff_user(),
-            'fiber_page': other_root,
-        })
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul>'
-                 '<li class="home first last">'
-                 '<a href="/">home</a>'
-                 '</li>'
-                 '</ul>'))
-
-    def test_show_admin_menu_all(self):
-        # render menu with all pages
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_menu 'main' 1 999 "all" %}
-            """
-        )
-        c = Context({
-            'user': self.get_staff_user(),
-            'fiber_page': Page.objects.get_by_url('/'),
-        })
-
-        with self.assertNumQueries(2):
-            self.assertEquals(
-                condense_html_whitespace(t.render(c)),
-                ('<ul data-fiber-data=\'{ "type": "page", "add_url": "%(fiber_admin_page_add_url)s", "parent_id": 1 }\'>'
-                   '<li class="home first last">'
-                     '<a href="/" data-fiber-data=\'{ "can_edit": true, "type": "page", "id": 2, "parent_id": 1, "url": "%(fiber_admin_page_edit_url_home)s", "add_url": "%(fiber_admin_page_add_url)s" }\'>home</a>'
-                     '<ul data-fiber-data=\'{ "type": "page", "add_url": "%(fiber_admin_page_add_url)s", "parent_id": 2 }\'>'
-                       '<li class="section1 first">'
-                         '<a href="/section1/" data-fiber-data=\'{ "can_edit": true, "type": "page", "id": 3, "parent_id": 2, "url": "%(fiber_admin_page_edit_url_section1)s", "add_url": "%(fiber_admin_page_add_url)s" }\'>section1</a>'
-                         '<ul data-fiber-data=\'{ "type": "page", "add_url": "%(fiber_admin_page_add_url)s", "parent_id": 3 }\'>'
-                           '<li class="sub1 first">'
-                             '<a href="/section1/sub1/" data-fiber-data=\'{ "can_edit": true, "type": "page", "id": 5, "parent_id": 3, "url": "%(fiber_admin_page_edit_url_sub1)s", "add_url": "%(fiber_admin_page_add_url)s" }\'>sub1</a>'
-                           '</li>'
-                           '<li class="sub2 last">'
-                             '<a href="/section1/sub2/" data-fiber-data=\'{ "can_edit": true, "type": "page", "id": 6, "parent_id": 3, "url": "%(fiber_admin_page_edit_url_sub2)s", "add_url": "%(fiber_admin_page_add_url)s" }\'>sub2</a>'
-                           '</li>'
-                         '</ul>'
-                       '</li>'
-                       '<li class="section2 last">'
-                         '<a href="/section2/" data-fiber-data=\'{ "can_edit": true, "type": "page", "id": 4, "parent_id": 2, "url": "%(fiber_admin_page_edit_url_section2)s", "add_url": "%(fiber_admin_page_add_url)s" }\'>section2</a>'
-                       '</li>'
-                     '</ul>'
-                   '</li>'
-                 '</ul>' % {
-                        'fiber_admin_page_add_url': reverse('fiber_admin:fiber_page_add'),
-                        'fiber_admin_page_edit_url_home': reverse('fiber_admin:fiber_page_change', args=(2, )),
-                        'fiber_admin_page_edit_url_section1': reverse('fiber_admin:fiber_page_change', args=(3, )),
-                        'fiber_admin_page_edit_url_section2': reverse('fiber_admin:fiber_page_change', args=(4, )),
-                        'fiber_admin_page_edit_url_sub1': reverse('fiber_admin:fiber_page_change', args=(5, )),
-                        'fiber_admin_page_edit_url_sub2': reverse('fiber_admin:fiber_page_change', args=(6, )),
-                        }
-                 ))
-
-    def test_show_page_content(self):
-        # The show_page_content templatetag should support rendering content from multiple pages in one view.
-
-        p1 = Page.objects.create(title='p1')
-        c1 = ContentItem.objects.create(content_html='<p>c1</p>')
-        PageContentItem.objects.create(content_item=c1, page=p1, block_name='test_block')
-        p2 = Page.objects.create(title='p2')
-        c2 = ContentItem.objects.create(content_html='<p>c2</p>')
-        PageContentItem.objects.create(content_item=c2, page=p2, block_name='test_block')
-
-        t = Template("""
-            {% load fiber_tags %}
-            {% show_page_content second_page 'test_block' %}
-            {% show_page_content 'test_block' %}
-            """
-            )
-
-        c = Context({
-            'fiber_page': p1,
-            'second_page': p2
-        })
-
-        self.assertEquals(
-            condense_html_whitespace(t.render(c)),
-            ('<div><div class="content"><p>c2</p></div></div><div><div class="content"><p>c1</p></div></div>'))
-
-
-class TestView(View):
-    pass
-
-
-urlpatterns = patterns('',
-    url(r'^test_url/$', TestView.as_view(), name='another_named_url'),
-)
-
-
-class TestUtilsURLValidator(TestCase):
-    urls = urlpatterns  # use this url conf for our tests
-    validator = FiberURLValidator()
-
-    def test_passes_normal(self):
-        self.assertEqual(self.validator('http://www.google.com/'), None)
-
-    def test_passes_url_contains_anchor(self):
-        self.assertEqual(self.validator('/some/page/#SomeAnchor'), None)
-
-    def test_passes_url_contains_querystring_and_anchor(self):
-        self.assertEqual(self.validator('/some/page/?id=12&user_id=1#SomeAnchor'), None)
-
-    def test_named_url(self):
-        # Must raise if named url does not exist
-        with self.assertRaises(ValidationError):
-            self.validator('"some_named_url"')
-
-        # Named url does exist
-        self.assertEquals(self.validator('"another_named_url"'), None)
-
-        # A fiber page also uses that named_url
-        Page.objects.create(title='some_page', url='"another_named_url"').save()
-        self.assertEquals(self.validator('"another_named_url"'), None)
-
-
-class TestFiberPageMixin(TestCase):
-
-    def setUp(self):
-        """
-        Generate test data
-        """
-        self.a = Page.objects.create(title='a')
-        self.aa = Page.objects.create(title='aa', parent=self.a, url='aa')
-        Page.objects.create(title='ab', parent=self.a, url='ab')
-        self.aaa = Page.objects.create(title='aaa', parent=self.aa, url='aaa')
-
-    def test_current_pages(self):
-        """
-        `get_fiber_current_pages` must return the correct list of pages that are marked as
-        current.
-        """
-        class TestView(FiberPageMixin, View):
-            def get_fiber_page_url(self):  # FiberPageMixin requires this method
-                # mock a request
-                return self.url
-
-        view = TestView()
-        view.url = 'aa'
-        # There should be no root node, but we do want a page; aa
-        self.assertEqual([self.aa], view.get_fiber_current_pages())
-
-        view = TestView()
-        view.url = 'aaa'
-        # again no root node but we do want to pages; aa -> aaa
-        self.assertEqual([self.aa, self.aaa], view.get_fiber_current_pages())
-
-    # TODO write tests for `mark_current_regexes` behavior
-
-
 class TestContentItem(TestCase):
     def test_unicode(self):
         # with name
@@ -693,35 +363,6 @@ class TestContentItem(TestCase):
         # - load contentitem
         content_item1 = ContentItem.objects.get(id=content_item1.id)
         self.assertEqual(content_item1.used_on_pages_data, [dict(url='/abc/', title='p1')])
-
-class TestSitemap(TestCase):
-
-    def setUp(self):
-        """
-        Generate test data.
-        """
-        self.a = Page.objects.create(title='a')
-        self.aa = Page.objects.create(title='aa', parent=self.a, url='aa')
-
-    def test_sitemap_class(self):
-        """
-        Sitemap class should list the 2 test pages.
-        """
-
-        urls = FiberSitemap().get_urls()
-        self.assertEqual(len(urls), 2)
-
-    def test_public(self):
-        """
-        Only public pages go in the sitemap.
-        """
-
-        self.a.is_public = False
-        self.a.save()
-
-        urls = FiberSitemap().get_urls()
-        self.assertEqual(len(urls), 1)
-
 
 
 def mock_tz_now():
